add user settings

+++ src/SettingsManger.php
@@ -9,21 +9,21 @@ class SettingsManger
 {
     public function all()
     {
-        return Setting::pluck('value', 'key');
+        return Setting::whereNull('owner_id')->pluck('value', 'key');
     }
 
     public function set($key, $value)
     {
-        Setting::updateOrCreate(['key' => $key], ['value' => $value]);
+        Setting::updateOrCreate(['key' => $key, 'owner_id' => null], ['value' => $value]);
     }
 
     public function get($key, $default = null)
     {
-        $cache_key = config('settings.cache_prefix') . $key;
+        $cache_key = config('settings.cache_prefix') . $key . '_global';
         $duration = config('settings.cache_duration');
 
         $value = Cache::remember($cache_key, $duration, function () use ($key) {
-            return Setting::where('key', $key)->pluck('value')->first();
+            return Setting::where('key', $key)->whereNull('owner_id')->pluck('value')->first();
         });
 
         return $value ? $value : $default;
@@ -31,12 +31,12 @@ class SettingsManger
 
     public function forget($key)
     {
-        Setting::where('key', $key)->first()->delete();
+        Setting::where('key', $key)->whereNull('owner_id')->first()->delete();
     }
 
     public function flush()
     {
-        Setting::get()->each(function ($item) {
+        Setting::whereNull('owner_id')->each(function ($item) {
             $item->delete();
         });
     }
